agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
copilotkit==0.1.44
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.12
forbiddenfruit==0.1.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.14.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.24
langchain-core==0.3.56
langchain-mcp-adapters==0.0.11
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-api==0.1.18
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.7
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.7
langgraph-sdk==0.1.63
langsmith==0.3.37
matplotlib-inline==0.1.7
mcp==1.8.0
nest-asyncio==1.6.0
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
partialjson==0.0.8
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
